Add background colors to form section
I'm seeing that it might be useful to further break branches down
to certain areas as well, instead of doing all the styles just in
one branch. This branch is called "form styles" but I've been
styling the entire page. In addition to making the commits easier
to wrangle it also keeps me from having to specify areas in commit
messages.